# Conflicts:
#	src/main/java/ru/geekbrains/domodel/services/core/NewsServiceImpl.java

+++ src/main/java/ru/geekbrains/domodel/entities/Account.java
@@ -38,13 +38,13 @@ public class Account {
 
     // Количество соток участка
     @Column(name = "acres_amount")
-    private Integer acresAmount;
+    private Double acresAmount;
 
     // Список счетчиков данного аккаунта
     @OneToMany(mappedBy = "account")
     private List<Meter> meters = new ArrayList<>();
 
-    // Список счетов, выставленных для данного аккаунта
+    // Список платежных документов (счетов), выставленных для данного аккаунта
     @OneToMany(mappedBy = "account")
     private List<Bill> bills = new ArrayList<>();
 }
